+++ packages/pulumi-aws/src/apps/admin/createAdminPulumiApp.ts
@@ -1,4 +1,4 @@
-import { PulumiAppParamCallback } from "@webiny/pulumi";
+import { PulumiAppParam, PulumiAppParamCallback } from "@webiny/pulumi";
 import { createReactPulumiApp, CustomDomainParams } from "~/apps";
 
 export type AdminPulumiApp = ReturnType<typeof createReactPulumiApp>;
@@ -12,6 +12,11 @@ export interface CreateAdminPulumiAppParams {
      * or add additional ones into the mix.
      */
     pulumi?: (app: AdminPulumiApp) => void | Promise<void>;
+
+    /**
+     * Prefixes names of all Pulumi cloud infrastructure resource with given prefix.
+     */
+    pulumiResourceNamePrefix?: PulumiAppParam<string>;
 }
 
 export const createAdminPulumiApp = (projectAppParams: CreateAdminPulumiAppParams) => {